Move webpack files to assets dir
