+++ assets/src/__tests__/fixtures/version_from_scratch.json
