+++ assets/src/__tests__/fixtures/version_with_complete_searches.json
