+++ assets/src/__tests__/fixtures/version_with_no_searches.json
