+++ assets/src/__tests__/fixtures/version_with_some_searches.json
